🎨 Spread helper functions in separate files
